refactor(pkg): use pkg directory
move folders inside pkg directory per https://github.com/golang-standards/project-layout
