+++ cmd/upload.go
@@ -23,9 +23,9 @@ import (
 	"github.com/storacha/go-ucanto/core/result"
 	"github.com/storacha/go-ucanto/did"
 	"github.com/storacha/go-ucanto/principal"
-	"github.com/storacha/guppy/car/sharding"
-	"github.com/storacha/guppy/client"
 	"github.com/storacha/guppy/cmd/util"
+	"github.com/storacha/guppy/pkg/car/sharding"
+	"github.com/storacha/guppy/pkg/client"
 	"github.com/urfave/cli/v2"
 )
