+++ pkg/capability/uploadadd/result.go
