+++ pkg/capability/uploadlist/result.go
